requirements.txt
langchain
langgraph
langchain-google-genai
langchain-core
langchain-community
faiss-cpu
pydantic
python-dotenv
